test-bot: split into multiple files, add RSpec tests.
This should make it easier to add additional (particularly regression)
tests and generally refactor Homebrew/homebrew-test-bot.
