+++ .travis.yml
@@ -1,3 +1,6 @@
+# don't try to be clever with the Gemfile
+language: c
+
 branches:
   only:
     - master
