+++ spec/spec_helper.rb
@@ -1,0 +1,38 @@
+# frozen_string_literal: true
+
+require "simplecov"
+SimpleCov.start do
+  add_filter "/vendor/"
+  minimum_coverage 10
+end
+
+PROJECT_ROOT ||= Pathname(__dir__).parent
+STUB_PATH ||= PROJECT_ROOT/"spec/stub"
+$LOAD_PATH.unshift(STUB_PATH)
+
+Dir.glob("#{PROJECT_ROOT}/lib/**/*.rb").each do |file|
+  next if file.include?("/extend/os/")
+
+  require file
+end
+
+require "global"
+
+SimpleCov.formatters = [SimpleCov::Formatter::HTMLFormatter]
+
+require "bundler"
+require "rspec/support/object_formatter"
+
+RSpec.configure do |config|
+  config.filter_run_when_matching :focus
+  config.expect_with :rspec do |c|
+    c.max_formatted_output_length = 200
+  end
+
+  # Never truncate output objects.
+  RSpec::Support::ObjectFormatter.default_instance.max_formatted_output_length = nil
+
+  config.around do |example|
+    Bundler.with_clean_env { example.run }
+  end
+end
